requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
SpeechRecognition==3.14.1
typing_extensions==4.12.2
urllib3==2.3.0
